% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_QRSIDE}
\alias{run_QRSIDE}
\title{执行 QR-SIDE 解卷积}
\usage{
run_QRSIDE(
  sp_counts,
  pos,
  markerframe,
  Num_topic,
  Num_HVG,
  dim_embed,
  max_marker,
  markerflag
)
}
\arguments{
\item{sp_counts}{空间表达矩阵}

\item{pos}{位置坐标}

\item{markerframe}{标记基因框架}

\item{Num_topic}{主题数}

\item{Num_HVG}{高变基因数}

\item{dim_embed}{嵌入维度}

\item{max_marker}{最大标记基因数}

\item{markerflag}{标记基因标志}
}
\value{
解卷积结果
}
\description{
主函数，用于空间转录组数据的细胞类型解卷积分析
}
